Updated Zoom Readme and made some refactorings to solution
