+++ Bot.Builder.Community.Samples.sln
@@ -1,12 +1,7 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29318.209
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BestMatch Middleware Sample", "samples\BestMatch Middleware Sample\BestMatch Middleware Sample.csproj", "{7927A7EE-C327-4403-A5F4-B5E2C0BF4D47}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alexa Adapter Sample", "samples\Alexa Adapter Sample\Alexa Adapter Sample.csproj", "{8183FE0F-6F69-48B4-B2C3-7FEB48DCECD6}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Location Dialog Sample", "samples\Location Dialog Sample\Location Dialog Sample.csproj", "{10BEDE11-C9DE-4B67-9467-4E48272229C6}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google Adapter Sample", "samples\Google Adapter Sample\Google Adapter Sample.csproj", "{3312C17E-01FA-4390-9B8F-6E7733E0B1E7}"
@@ -23,11 +18,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sentiment Middleware Sample
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChoiceFlow Dialog Sample", "samples\ChoiceFlow Dialog Sample\ChoiceFlow Dialog Sample.csproj", "{D1C3490D-477A-4D79-A41C-4D15CA9DC183}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BestMatch Middleware Sample", "samples\BestMatch Middleware Sample\BestMatch Middleware Sample.csproj", "{8E97F66B-783C-49BC-8625-914EB4D4B3A8}"
-End Project
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BestMatch Middleware Sample", "samples\BestMatch Middleware Sample\BestMatch Middleware Sample.csproj", "{8E97F66B-783C-49BC-8625-914EB4D4B3A8}"
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RingCentral Adapter Sample", "samples\RingCentral Adapter Sample\RingCentral Adapter Sample.csproj", "{231952BC-62A6-4D86-82FB-A2A5940EF403}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alexa Adapter Sample", "samples\Alexa Adapter Sample\Alexa Adapter Sample.csproj", "{53BCE929-C57D-4AF9-91C9-FCA5C7E2CE41}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Zoom Adapter Sample", "samples\Zoom Adapter Sample\Zoom Adapter Sample.csproj", "{D5F5BFB0-CDEC-4980-9EBB-864724B6B2B6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,14 +32,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8183FE0F-6F69-48B4-B2C3-7FEB48DCECD6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8183FE0F-6F69-48B4-B2C3-7FEB48DCECD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8183FE0F-6F69-48B4-B2C3-7FEB48DCECD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8183FE0F-6F69-48B4-B2C3-7FEB48DCECD6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7927A7EE-C327-4403-A5F4-B5E2C0BF4D47}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7927A7EE-C327-4403-A5F4-B5E2C0BF4D47}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7927A7EE-C327-4403-A5F4-B5E2C0BF4D47}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7927A7EE-C327-4403-A5F4-B5E2C0BF4D47}.Release|Any CPU.Build.0 = Release|Any CPU
 		{10BEDE11-C9DE-4B67-9467-4E48272229C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{10BEDE11-C9DE-4B67-9467-4E48272229C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{10BEDE11-C9DE-4B67-9467-4E48272229C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -87,6 +76,10 @@ Global
 		{53BCE929-C57D-4AF9-91C9-FCA5C7E2CE41}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{53BCE929-C57D-4AF9-91C9-FCA5C7E2CE41}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{53BCE929-C57D-4AF9-91C9-FCA5C7E2CE41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5F5BFB0-CDEC-4980-9EBB-864724B6B2B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5F5BFB0-CDEC-4980-9EBB-864724B6B2B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5F5BFB0-CDEC-4980-9EBB-864724B6B2B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5F5BFB0-CDEC-4980-9EBB-864724B6B2B6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
