+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/BodyItem.cs
@@ -1,0 +1,6 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class BodyItem
+    {
+    }
+}
