+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/InteractiveMessageActionsPayload.cs
@@ -1,0 +1,7 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class InteractiveMessageActionsPayload : InteractiveMessagePayload
+    {
+        public ZoomAction ActionItem { get; set; }
+    }
+}
