+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/InteractiveMessagePayload.cs
@@ -1,0 +1,9 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class InteractiveMessagePayload : Payload
+    {
+        public string MessageId { get; set; }
+
+        public ChatResponseContent Original { get; set; }
+    }
+}
