+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/MessageBodyItemWithLink.cs
@@ -1,0 +1,7 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class MessageBodyItemWithLink : MessageItem
+    {
+        public string Link { get; set; }
+    }
+}
