+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/MessageItem.cs
@@ -1,0 +1,11 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class MessageItem : BodyItem
+    {
+        public string Type => "message";
+
+        public string Text { get; set; }
+
+        public Style Style { get; set; }
+    }
+}
