+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/MessageItemWithEditableText.cs
@@ -1,0 +1,7 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class MessageItemWithEditableText : MessageItem
+    {
+        public bool Editable { get; set; } = true;
+    }
+}
