+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/Style.cs
@@ -1,0 +1,9 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class Style
+    {
+        public string Color { get; set; }
+        public bool Bold { get; set; }
+        public bool Italic { get; set; }
+    }
+}
