+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/SubHead.cs
@@ -1,0 +1,9 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class SubHead
+    {
+        public string Text { get; set; }
+
+        public Style Style { get; set; }
+    }
+}
