+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/ZoomAction.cs
@@ -1,0 +1,9 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class ZoomAction
+    {
+        public string Text { get; set; }
+        public string Value { get; set; }
+        public string Style { get; set; }
+    }
+}
