+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/ZoomAttachmentInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class ZoomAttachmentInfo
+    {
+        public ZoomAttachmentInfoContent Title { get; set; }
+
+        public ZoomAttachmentInfoContent Description { get; set; }
+    }
+}
