+++ libraries/Bot.Builder.Community.Adapters.Zoom/Models/ZoomAttachmentInfoContent.cs
@@ -1,0 +1,8 @@
+﻿namespace Bot.Builder.Community.Adapters.Zoom.Models
+{
+    public class ZoomAttachmentInfoContent
+    {
+        public string Text { get; set; }
+        public Style Style { get; set; }
+    }
+}
